Added Spring Security features to API

+++ src/main/java/com/spoctexter/userProfile/UserProfileController.java
@@ -1,11 +1,21 @@
 package com.spoctexter.userProfile;
 
+import com.spoctexter.exception.ForbiddenAccessException;
+import com.spoctexter.jwt.JwtConfig;
 import com.sun.istack.NotNull;
+import io.jsonwebtoken.Jwts;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.parameters.P;
 import org.springframework.web.bind.annotation.*;
 
+import javax.crypto.SecretKey;
+import javax.servlet.http.HttpServletResponse;
 import javax.validation.Valid;
+import java.security.Principal;
+import java.time.LocalDate;
 import java.time.OffsetDateTime;
+import java.util.Date;
 import java.util.UUID;
 
 @RestController
@@ -13,10 +23,17 @@ import java.util.UUID;
 public class UserProfileController {
 
     private final UserProfileService userService;
+    private final JwtConfig jwtConfig;
+    private final SecretKey secretKey;
 
     @Autowired
-    public UserProfileController(UserProfileService userService) {
+    public UserProfileController(UserProfileService userService,
+                                 JwtConfig jwtConfig,
+                                 SecretKey secretKey
+    ) {
         this.userService = userService;
+        this.jwtConfig = jwtConfig;
+        this.secretKey = secretKey;
     }
 
     @PostMapping(path = "addUser") //Another way to add a user profile
@@ -48,33 +65,58 @@ public class UserProfileController {
     }
 
     @GetMapping(path = "id={id}")
-    public UserProfile getUserProfileByID(@NotNull @PathVariable("id") UUID id) {
-        return userService.getUserProfileByID(id);
+    public UserProfile getUserProfileByID(@NotNull @PathVariable("id") UUID id, Principal principal) {
+        if(principal.getName().equals(userService.getUserProfileByID(id).getUserAccount().getUsername())) {
+            return userService.getUserProfileByID(id);
+        } else {
+            throw new ForbiddenAccessException("You do not have access to this resource.");
+        }
     }
 
     @GetMapping(path = "email={email}")
-    public UserProfile getUserProfileByEmail(@NotNull @PathVariable("email") String email) {
-        return userService.getUserProfileByEmail(email);
+    public UserProfile getUserProfileByEmail(@NotNull @PathVariable("email") String email, Principal principal) {
+        if(principal.getName().equals(userService.getUserProfileByEmail(email).getUserAccount().getUsername())) {
+            return userService.getUserProfileByEmail(email);
+        } else {
+            throw new ForbiddenAccessException("You do not have access to this resource hahaha.");
+        }
+
     }
 
     @GetMapping(path = "phoneNumber={phoneNumber}")
-    public UserProfile getUserProfileByPhoneNumber(@NotNull @PathVariable("phoneNumber") String phoneNumber) {
-        return userService.getUserProfileByPhoneNumber(phoneNumber);
+    public UserProfile getUserProfileByPhoneNumber(@NotNull @PathVariable("phoneNumber") String phoneNumber, Principal principal) {
+        if (principal.getName().equals(userService.getUserProfileByPhoneNumber(phoneNumber).getUserAccount().getUsername())) {
+            return userService.getUserProfileByPhoneNumber(phoneNumber);
+        } else {
+            throw new ForbiddenAccessException("You do not have access to this resource.");
+        }
     }
 
     @DeleteMapping(path = "id={id}")
-    public void deleteUserProfileById(@NotNull @PathVariable("id") UUID id) {
-        userService.deleteUserProfileByID(id);
+    public void deleteUserProfileById(@NotNull @PathVariable("id") UUID id, Principal principal) {
+        if(principal.getName().equals(userService.getUserProfileByID(id).getUserAccount().getUsername())) {
+            userService.deleteUserProfileByID(id);
+        } else {
+            throw new ForbiddenAccessException("You do not have access to this resource.");
+        }
     }
 
     @DeleteMapping(path = "email={email}")
-    public void deleteUserProfileByEmail(@NotNull @PathVariable("email") String email) {
-        userService.deleteUserProfileByEmail(email);
+    public void deleteUserProfileByEmail(@NotNull @PathVariable("email") String email, Principal principal) {
+        if(principal.getName().equals(userService.getUserProfileByEmail(email).getUserAccount().getUsername())) {
+            userService.deleteUserProfileByEmail(email);
+        } else {
+            throw new ForbiddenAccessException("You do not have access to this resource.");
+        }
     }
 
     @DeleteMapping(path = "phone={phoneNumber}")
-    public void deleteUserProfileByPhoneNumber(@NotNull @PathVariable("phoneNumber") String phoneNumber) {
-        userService.deleteUserProfileByPhoneNumber(phoneNumber);
+    public void deleteUserProfileByPhoneNumber(@NotNull @PathVariable("phoneNumber") String phoneNumber, Principal principal) {
+        if (principal.getName().equals(userService.getUserProfileByPhoneNumber(phoneNumber).getUserAccount().getUsername())) {
+            userService.deleteUserProfileByPhoneNumber(phoneNumber);
+        } else {
+            throw new ForbiddenAccessException("You do not have access to this resource.");
+        }
     }
 
     @PutMapping(path = "update")
@@ -83,39 +125,35 @@ public class UserProfileController {
             @RequestParam (name = "newPhoneNumber") String newPhoneNumber,
             @RequestParam (name = "newFirstName") String newFirstName,
             @RequestParam (name = "newLastName") String newLastName,
-            @RequestParam (name = "newUserName") String newUserName,
-            @RequestParam (name = "currentEmail") String currentEmail
+            @RequestParam (name = "newUsername") String newUsername,
+            @RequestParam (name = "currentEmail") String currentEmail,
+            Authentication authentication,
+            Principal principal,
+            HttpServletResponse response
     ) {
-        userService.updateUserProfile(newEmail, newPhoneNumber, newFirstName, newLastName, newUserName, currentEmail);
+        if( userService.getUserProfileByEmail(currentEmail).getUserAccount().getUsername().equals(principal.getName()) ){
+            userService.updateUserProfile(newEmail, newPhoneNumber, newFirstName, newLastName, newUsername, currentEmail);
+
+            String subject;
+
+            if (newUsername.equals("")) {
+                subject = principal.getName();
+            } else {
+                subject = newUsername;
+            }
+
+            String token = Jwts.builder()
+                    .setSubject(subject)
+                    .claim("authorities", authentication.getAuthorities())
+                    .setIssuedAt(new Date())
+                    .setExpiration(java.sql.Date.valueOf(LocalDate.now().plusDays(jwtConfig.getTokenExpirationAfterDays())))
+                    .signWith(secretKey)
+                    .compact();
+
+            response.addHeader(jwtConfig.getAuthorizationHeader(), jwtConfig.getTokenPrefix() + token);
+
+        } else {
+            throw new ForbiddenAccessException("You do not have access to this resource.");
+        }
     }
-
-
-    /*@Deprecated
-    @PutMapping(path = "oldEmail={email}/newEmail={newEmail}")
-    public void updateUserProfileEmail(@NotNull @PathVariable("email") String email, @NotNull @PathVariable("newEmail") String newEmail) {
-        userService.updateUserProfileEmail(email, newEmail);
-    }
-
-    @PutMapping(path = "oldPhone={phone}/newPhone={newPhone}")
-    public void updateUserProfilePhone(@NotNull @PathVariable("phone") String phoneNumber, @NotNull @PathVariable("newPhone") String newPhoneNumber) {
-        userService.updateUserProfilePhoneNumber(phoneNumber, newPhoneNumber);
-    }
-
-    @PutMapping(path = "email={email}/newFirstName={newFirstName}")
-    public void updateUserProfileFirstName(
-            @NotNull @PathVariable("email") String email,
-            @NotNull @PathVariable("newFirstName") String newFirstName) {
-
-        userService.updateUserProfileFirstName(email, newFirstName);
-    }
-
-    @PutMapping(path = "email={email}/newLastName={newLastName}")
-    public void updateUserProfileLastName(
-            @NotNull @PathVariable("email") String email,
-            @NotNull @PathVariable("newLastName") String newLastName) {
-
-        userService.updateUserProfileLastName(email, newLastName);
-    }*/
-
-
 }
